1) Implemented so-called "Victory types". 2) Some neater formatting. 3) Fix: show the first row of the table when in desktop mode.
Victory Types: 7 Wonders Duel has 3 victory tracks: Military, Scientific or Points. When a player has won the military or science track, the points of neither player matter anymore, so I am hiding those. (If a player for some reason still wants to enter points, they can, but they have to wait clicking the victory track until last.)

There were different ways in which I could have set the victoryType value, but I decided to do it at the level of the top component, because there are a couple of other things I need to do there while I am at it. Maybe not stylistically pure, but it sure adds to readability if you do not have to keep ducking in and out of components.