1) Adding the local storage name of the games log. 2) Upon "calling the game", write the games log to local storage. 3) Upon loading the app, retrieve the game log from the local storage. 4) Display only the 10 most recent items (currently: 3, because that is easier to test).
I don't like the hackish way some of these things are achieved, I am constantly wrestling with getting a non-reactive copy from the reactive gamesLog, I will have to look at this again later.

The way I deal with the local storage name is perhaps a bit weird, but in Angular these sort of constants appear to be a Big Thing and this seems to work, so why not?