1) Renamed 'URLs' to 'localStorageLib' because I added a helper function. 2) Added a "more items" button to the games log. Set the maximum amount of items to 6 instead of 10 accordingly, as the user now has to option to view more items at their own discretion. 3) Niceified the games log output.
In order to share one of the niceification functions, I had to move it to its own file. As I already had a file that had to do with local storage, I decided to use it for both goals. Whether this is a good thing or not I do not know. I guess it does not matter much as long as the included file is small.