1) Added comments to functions. 2) Improved validation functions. 3) Some validation functions were called after a value had been changed. I fixed this, although I cannot guarantee this will work equally in all browsers.
Re function comments: I am not sure if these are the right type of comments. I have added them because I believe some form of comment is always better than none. My doubt stems from the fact that often in software the flow of a program can be deduced from reading the function interfaces. However, in a reactive framework where functions change reactive values, that flow remains largely hidden. Is there a way of commenting that highlights the meaning of a function in reactive frameworks, or should I write entirely different types of functions? (I say 'reactive frameworks', but I imagine the same forces are at play in other types of MVC or MVVM systems.)