1) Fix: store and display games log items in the right order. 2) Revert paths in index.html to what I believe were the original values.
Re #1: after running a number of tests, I noticed that the games log got saved as: old items reversed + new item last. (The intended order is from old to new, so the "new item last" was as it should be.

As there is really only one place where the sort order of games log item gets reversed, namely in the computed value sortedItemList (components/games-log.vue), that seemed to be the spot to look. Turns out the method I used, reverse(), changes the array it is a method of. I had to used toReversed() instead.

Re #2: I changed these values because it helped with local development, but I am not sure this was the best way, so I changed them back.